Rename PROFILED to _ENV_PROFILED and prevent its inheritance through sudo.
Previously, PROFILED would be set to 'true' -- this caused problems when
invoking a shell via sudo, for which a full reinitialization is needed to
make sure e.g. /hpc/settings.sh is loaded properly. Now, _ENV_PROFILED is
set to $USER -- if another shell starts up under another user, it can see
that the value is not what is expected and can therefore rebuild the
shell environment from scratch.